Add authentication with role-based access
