+++ README.md
@@ -25,6 +25,7 @@ npm run dev
 - Сүрэг: үүсгэх, азарга сонгох, гишүүдийн тоо
 - Хайлт: ерөнхий (ямар ч талбар), болон **сүрэг дотор** хайлт
 - Пагинац: ерөнхий жагсаалт, сүрэг дотор
+- Нэвтрэлт: админ ба хэрэглэгчийн эрхээр нэвтрэх, бүртгэл үүсгэх
 
 ## Тохиргоо
-- `server/.env` дотор Mongo URI ба PORT тодорхойлсон.
+- `server/.env` дотор Mongo URI, PORT болон JWT_SECRET тодорхойлно.
