+++ server/models/User.js
@@ -1,0 +1,13 @@
+import mongoose from "mongoose";
+
+const userSchema = new mongoose.Schema(
+  {
+    name: { type: String, required: true, trim: true },
+    email: { type: String, required: true, unique: true, lowercase: true, trim: true },
+    password: { type: String, required: true },
+    role: { type: String, enum: ["admin", "customer"], default: "customer" }
+  },
+  { timestamps: true }
+);
+
+export default mongoose.model("User", userSchema);
