+++ server/server.js
@@ -5,6 +5,8 @@ import cors from "cors";
 import morgan from "morgan";
 import horsesRouter from "./routes/horses.js";
 import herdsRouter from "./routes/herds.js";
+import authRouter from "./routes/auth.js";
+import { authenticate } from "./middlewares/auth.js";
 
 const app = express();
 app.use(cors());
@@ -13,8 +15,9 @@ app.use(morgan("dev"));
 
 app.get("/", (_req, res) => res.send("Horse Registry API ажиллаж байна"));
 
-app.use("/api/horses", horsesRouter);
-app.use("/api/herds", herdsRouter);
+app.use("/api/auth", authRouter);
+app.use("/api/horses", authenticate, horsesRouter);
+app.use("/api/herds", authenticate, herdsRouter);
 
 const PORT = process.env.PORT || 4000;
 mongoose
