+++ server/services/token.js
@@ -1,0 +1,8 @@
+import jwt from "jsonwebtoken";
+
+const JWT_SECRET = process.env.JWT_SECRET || "dev-secret-change-me";
+const TOKEN_EXPIRES_IN = "7d";
+
+export const signToken = (payload) => jwt.sign(payload, JWT_SECRET, { expiresIn: TOKEN_EXPIRES_IN });
+
+export const verifyToken = (token) => jwt.verify(token, JWT_SECRET);
